NTI-10700 - clean up sortable library flags
