--- src/collection-sortable/components/Grid.jsx
@@ -1,3 +1,0 @@
-import { Layouts } from '@nti/web-commons';
-
-export const Grid = Layouts.grid(242, 14);
