+++ src/collection/components/Grid.jsx
@@ -1,3 +1,3 @@
 import { Layouts } from '@nti/web-commons';
 
-export default Layouts.grid(242, 14);
+export const Grid = Layouts.grid(242, 14);
