+++ src/collection/components/Group.jsx
@@ -1,12 +1,11 @@
-import React from 'react';
 import PropTypes from 'prop-types';
 
 import { Errors } from '@nti/web-commons';
 
 import Card from '../../card/View';
 
-import Grid from './Grid';
-import GroupHeader from './GroupHeader';
+import { Grid } from './Grid';
+import { GroupHeader } from './GroupHeader';
 
 const Section = styled('section')`
 	margin-bottom: 2rem;
@@ -28,12 +27,14 @@ CourseCollectionGroup.propTypes = {
 
 	getSectionTitle: PropTypes.func,
 	onCourseDelete: PropTypes.func,
+	onSortChange: PropTypes.func,
 };
-export default function CourseCollectionGroup({
+export function CourseCollectionGroup({
 	group,
 	mobile,
 	getSectionTitle,
 	onCourseDelete,
+	onSortChange,
 }) {
 	if (group.error) {
 		return <Errors.Message as="p" error={group.error} />;
@@ -49,15 +50,20 @@ export default function CourseCollectionGroup({
 				group={group}
 				mobile={mobile}
 				getSectionTitle={getSectionTitle}
+				onSortChange={onSortChange}
 			/>
 			<Grid as="ul">
-				{(columns) => (
+				{columns =>
 					group.Items.map(item => (
 						<li key={getKey(item)}>
-							<Card course={item} onDelete={onCourseDelete} variant={columns === 1 ? 'list-item' : 'card'} />
+							<Card
+								course={item}
+								onDelete={onCourseDelete}
+								variant={columns === 1 ? 'list-item' : 'card'}
+							/>
 						</li>
 					))
-				)}
+				}
 			</Grid>
 		</Section>
 	);
