+++ src/collection/utils/batch-generator.js
